Moved nos integration test from test functions to binary, split out in CI.
