+++ .circleci/config.yml
@@ -64,6 +64,15 @@ jobs:
         - checkout
         - run: cargo run --bin large_download
 
+    # Rust testing - NOS integration tests
+    # Run for all PR commits
+    rust_test_nosengine:
+      docker:
+        - image: kubos/kubos-dev:latest
+      steps:
+        - checkout
+        - run: cd test/integration/nosengine-rust && cargo run
+
     # Rust testing - Check formatting
     # Run for all PR commits
     rust_fmt:
@@ -144,6 +153,10 @@ workflows:
             branches:
               ignore: master
       - rust_test_large_upload:
+          filters:
+            branches:
+              ignore: master
+      - rust_test_nosengine:
           filters:
             branches:
               ignore: master
